Added Player username generation, writing and reading from saved playerprefs, also added template ui for whackamole minigame, to be developed upon later
