+++ Assets/Data Files/Adjectives.txt
@@ -1,0 +1,107 @@
+Adjective
+Adaptable
+Adventurous
+Affectionate
+Alert
+Ambitious
+Amiable
+Astute
+Attentive
+Authentic
+Aware
+Awesome
+Bold
+Brave
+Calm
+Capable
+Caring
+Compassionate
+Confident
+Considerate
+Consistant
+Courageous
+Courteous
+Curious
+Decisive
+Dependable
+Determined
+Dexterous
+Diligent
+Diplomatic
+Dynamic
+Earnest
+Encouraging
+Energetic
+Engaging
+Enthusiastic
+Fair
+Fearless
+Flexible
+Focused
+Forgiving
+Forthright
+Free-spirited
+Friendly
+Fun
+Fun-loving
+Generous
+Gentle
+Genuine
+Giving
+Graceful
+Gracious
+Happy
+Hardworking
+Honest
+Hopeful
+Humble
+Humorous
+Idealistic
+Imaginative
+Innovative
+Insightful
+Intuitive
+Inventive
+Joyful
+Just
+Kind
+Lively
+Loving
+Loyal
+Merry
+Motivated
+Motivational
+Nurturing
+Observant
+Open
+Open-hearted
+Open-minded
+Optimistic
+Organized
+Outgoing
+Patient
+Persistent
+Playful
+Positive
+Precise
+Punctual
+Purposeful
+Quick-witted
+Radiant
+Realistic
+Reflective
+Reliable
+Resourceful
+Sincere
+Sociable
+Social
+Strong
+Sympathetic
+Trusting
+Trustworthy
+Upbeat
+Vivacious
+Warm
+Welcoming
+Wise
+Zany
