+++ Assets/Data Files/CREDITS.txt
@@ -1,0 +1,7 @@
+Nouns list is from:
+http://supplies.thesmartteacher.com.s3.amazonaws.com/assets/exchange/list-of-nouns.pdf
+
+Adjectives list is from:
+https://goldarrowcamp.com/app/uploads/2020/08/positive-adjectives-to-describe-a-person-free-printable-poster.pdf
+
+(Let's hope numbers continue to stay in the public domain)
