+++ Assets/Prefabs/UI Components/Minigames/Card.prefab
@@ -261,14 +261,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 9067402012819106553}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbbea5cad4be7874992c8add62b220d5, type: 3}
+  m_Script: {fileID: 11500000, guid: 555055ad079f46a4bbfc4c9ef26e2221, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   index: 0
   shown: 0
   isQuestion: 0
   info: 
-  text: {fileID: 0}
+  text: {fileID: 685684580101491845}
 --- !u!114 &3134865276786979679
 MonoBehaviour:
   m_ObjectHideFlags: 0
