--- Assets/Scripts/Player/MinigameController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 444a8c09b70ef7d49a9842a0de6d5d3e
