--- Assets/saveManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9699e0d60bb4b3a4caffe399be14f444
